test: migrate integrations-github-patch to bdd
